Drop outdated VS 2008 support artifacts
These were added to support Python 2.7/VS 2008 on Windows. However they
have since been added to the `vs2008_express_vc_python_patch` package on
`conda-forge`, which is installed and run as part of the startup process
on AppVeyor as part of all feedstocks. As such this content is unused
and can be dropped.